replaced discovery, gateway and config apis with eureka and zuul apis, which now work correctly with security.
