--- config/src/main/java/com/qa/config/ConfigApplication.java
@@ -1,14 +1,0 @@
-package com.qa.config;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@SpringBootApplication
-@EnableConfigServer
-public class ConfigApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConfigApplication.class, args);
-	}
-}
