--- config/src/main/resources/application.properties
@@ -1,12 +1,0 @@
-# ----------------------------
-# config server settings
-# ----------------------------
-server.port=8081
-spring.application.name=config
-spring.cloud.config.server.git.uri=https://github.com/Perfspec/cv-app-config
-# ----------------------------
-# eureka connection settings
-# ----------------------------
-eureka.client.region = default
-eureka.client.registryFetchIntervalSeconds = 5
-eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
