--- config/src/test/java/com/qa/config/ConfigApplicationTests.java
@@ -1,16 +1,0 @@
-package com.qa.config;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ConfigApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
