--- discovery/src/main/java/com/qa/discovery/DiscoveryApplication.java
@@ -1,14 +1,0 @@
-package com.qa.discovery;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class DiscoveryApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DiscoveryApplication.class, args);
-	}
-}
