--- discovery/src/test/java/com/qa/discovery/DiscoveryApplicationTests.java
@@ -1,16 +1,0 @@
-package com.qa.discovery;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DiscoveryApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
