+++ eureka/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+server.port=8761
+
+eureka.client.register-with-eureka=false
+eureka.client.fetch-registry=true
