+++ eureka/src/test/java/com/qa/cv_manager/eureka/EurekaApplicationTests.java
@@ -1,0 +1,17 @@
+package com.qa.cv_manager.eureka;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class EurekaApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
+
