--- gateway/src/main/java/com/qa/gateway/constants/Constants.java
@@ -1,13 +1,0 @@
-package com.qa.gateway.constants;
-
-public class Constants {
-
-	private Constants() {}
-	
-	public static final String ALL = "*";
-	public static final String GET = "GET";
-	public static final String PUT = "PUT";
-	public static final String POST = "POST";
-	public static final String DELETE = "DELETE";
-	public static final String ALL_DIRS = "/**";
-}
