--- gateway/src/main/resources/bootstrap.properties
@@ -1,5 +1,0 @@
-spring.cloud.config.name=gateway
-spring.cloud.config.discovery.service-id=config
-spring.cloud.config.discovery.enabled=true
- 
-eureka.client.serviceUrl.defaultZone=http://localhost:8082/eureka/
