--- gateway/src/test/java/com/qa/gateway/GatewayApplicationTests.java
@@ -1,16 +1,0 @@
-package com.qa.gateway;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class GatewayApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
