+++ zuul/src/main/java/com/qa/cv_manager/zuul/ZuulApplication.java
@@ -1,0 +1,18 @@
+package com.qa.cv_manager.zuul;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+
+@SpringBootApplication
+@EnableZuulProxy
+@EnableDiscoveryClient
+public class ZuulApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZuulApplication.class, args);
+	}
+
+}
+
