+++ zuul/src/main/resources/application.yml
@@ -1,0 +1,2 @@
+zuul:
+  sensitiveHeaders: Cookie,Set-Cookie
