+++ zuul/src/test/java/com/qa/cv_manager/zuul/ZuulApplicationTests.java
@@ -1,0 +1,17 @@
+package com.qa.cv_manager.zuul;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class ZuulApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
+
